python-model/requirements.txt
fastapi
uvicorn[standard]
pillow
python-multipart